+++ test/integration/PaymentTypes/InvoiceSecuredTest.php
@@ -107,6 +107,46 @@ class InvoiceSecuredTest extends BaseIntegrationTest
     }
 
     /**
+     * Verify charge with Invoice Secured throws an error when invalid ip is set.
+     *
+     * @test
+     */
+    public function invoiceSecuredRequiresValidClientIpForCharge(): void
+    {
+        $clientIp = '123.456.789.123';
+        $this->unzer->setClientIp($clientIp);
+
+        /** @var InvoiceSecured $invoiceSecured */
+        $invoiceSecured = $this->unzer->createPaymentType(new InvoiceSecured());
+        $this->unzer->setClientIp(null); // Ensure that the invalid ip is only used for type creation.
+        $this->assertEquals($clientIp, $invoiceSecured->getGeoLocation()->getClientIp());
+
+        $customer = $this->getMaximumCustomer();
+        $customer->setShippingAddress($customer->getBillingAddress());
+        $basket = $this->createBasket();
+
+        $this->expectException(UnzerApiException::class);
+        $this->expectExceptionCode(ApiResponseCodes::CORE_INVALID_IP_NUMBER);
+
+        $invoiceSecured->charge(119.0, 'EUR', self::RETURN_URL, $customer, $basket->getOrderId(), null, $basket);
+    }
+
+    /**
+     * Verify creating a payment type with client ip header set, will overwrite the clientIp of the API resource.
+     *
+     * @test
+     */
+    public function verifySettingClientIpViaHeaderWillOverwriteClientIpOfTypeResource(): void
+    {
+        $clientIp = 'xxx.xxx.xxx.xxx';
+        $this->unzer->setClientIp($clientIp);
+
+        /** @var InvoiceSecured $invoiceSecured */
+        $invoiceSecured = $this->unzer->createPaymentType(new InvoiceSecured());
+        $this->assertEquals($clientIp, $invoiceSecured->getGeoLocation()->getClientIp());
+    }
+
+    /**
      * Verify Invoice Secured is chargeable.
      *
      * @test